Refactor: 메인페이지 컴포넌트화 코드 구조 간단하게 수정
