+++ src/pages/main/MainPage.module.css
@@ -1,53 +1,1 @@
-.container {
-  position: relative;
-  min-height: 100vh;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  background: url("/src/assets/img/common/bg.jpg") no-repeat center;
-  background-size: cover;
-}
 
-.header {
-  height: 16rem;
-  display: flex;
-  justify-content: flex-end;
-  @media (max-width: 767px) {
-    order: 1;
-    height: auto;
-  }
-}
-
-.link {
-  margin: 4.5rem 13rem 0 0;
-  z-index: 1;
-
-  @media screen and (min-width: 768px) and (max-width: 1023px) {
-    margin: 4.5rem 5rem 0 0;
-  }
-
-  @media (max-width: 767px) {
-    margin: 2.4rem auto;
-  }
-}
-
-.section {
-  width: 40rem;
-  height: 17.2rem;
-  margin: var(--spacing-24) auto 0;
-  order: 2;
-  z-index: 1;
-
-  @media (max-width: 767px) {
-    width: 30.5rem;
-    height: 15.6rem;
-    margin: 0 auto;
-  }
-}
-
-.lottie {
-  position: absolute;
-  min-width: 100vw;
-  bottom: 0;
-  order: 3;
-}
