+++ src/pages/main/components/IntroHeader.module.css
@@ -1,0 +1,22 @@
+.header {
+  height: 16rem;
+  display: flex;
+  justify-content: flex-end;
+  @media (max-width: 767px) {
+    order: 1;
+    height: auto;
+  }
+}
+
+.link {
+  margin: 4.5rem 13rem 0 0;
+  z-index: 1;
+
+  @media screen and (min-width: 768px) and (max-width: 1023px) {
+    margin: 4.5rem 5rem 0 0;
+  }
+
+  @media (max-width: 767px) {
+    margin: 2.4rem auto;
+  }
+}
