+++ src/pages/main/components/IntroLottie.module.css
@@ -1,0 +1,6 @@
+.lottie {
+  position: absolute;
+  min-width: 100vw;
+  bottom: 0;
+  order: 3;
+}
