+++ src/pages/main/components/IntroWrapper.jsx
@@ -1,0 +1,7 @@
+import styles from "./IntroWrapper.module.css";
+
+function IntroWrapper({ children }) {
+  return <div className={styles.container}>{children}</div>;
+}
+
+export default IntroWrapper;
